+++ POO_Java_Unicesumar/src/transportadoraAtv/TransporteNautico.java
@@ -2,7 +2,7 @@ package transportadoraAtv;
 
 public class TransporteNautico extends Transportadora{
 
-	double frete;
+	float frete;
 	
 	public TransporteNautico(int altura, int largura) {
 		super(altura, largura);
@@ -10,8 +10,8 @@ public class TransporteNautico extends Transportadora{
 	}
 
 	@Override
-	double calculadoraFrete() {
-		frete = (altura+largura) * 0.10;
+	float calculadoraFrete() {
+		frete = (float) ((altura+largura) * 0.10);
 		return frete;
