+++ POO_Java_Unicesumar/src/transportadoraAtv/TransporteTerreste.java
@@ -2,14 +2,14 @@ package transportadoraAtv;
 
 public class TransporteTerreste extends Transportadora {
 
-	double frete;
+	float frete;
 	public TransporteTerreste(int altura, int largura) {
 		super(altura, largura);
 	}
 
 	@Override
-	double calculadoraFrete() {
-		frete = (altura/largura) * 0.20;
+	float calculadoraFrete() {
+		frete = (float) ((altura/largura) * 0.20);
 		return frete;
 		
 	}
